fixed end points and patch api

+++ routers/menu.js
@@ -31,6 +31,12 @@ router.post('/', isAuthenticated, isAuthorized,check('data').notEmpty(),check('p
 
 
 router.delete('/:id', isAuthenticated, isAuthorized, check('id').notEmpty(), (req,res) => {
+
+    let errors = validationResult(req)
+    if(!errors.isEmpty()){
+        return res.status(400).json({errors: errors.array()})
+    }
+
     Menu.findByIdAndRemove(req.params.id)
         .then(() => res.status(201).send('Succesfully deleted menu'))
         .catch(() => res.status(500).send('Error deleted Menu'));
